Add android.test.base/stubs dependency
In preparation for the removal of the non-junit classes in the
android.test.base library from the android.jar this adds a dependency
on android.test.base/stubs to ensure this code will continue to
compile.

The following change descriptions were generated automatically and so
may be a little repetitive. They are provided to give the reviewer
enough information to check the comments match what has actually been
changed and check the reasoning behind the changes.

* Android.mk
    Added 'android.test.base.stubs' to LOCAL_JAVA_LIBRARIES because
    Browser2's will need them in order to compile once its classes are
    removed from the current SDK on which it currently depends.

Bug: 30188076
Test: make dist
Change-Id: I15823654c5cef130b759d82b80f325463f84e10c