[Browser2] Sync from chromium code, including increase target_sdk version to 28 (P) and enable cleartext traffic
Update Browser2 to match the SystemWebViewShell code in
the chromium tree.

Update Browser2 to target Android P (API level 28).
This is a part of the general effort to update apps to API level 28,
see crbug.com/898190 and b/115499280.

This patch also ensures that cleartextTrafficPermitted is set to
true to ensure backwards compatibility for the Browser2 app
(and to allow cleartext traffic, e.g. using unencrypted http
protocol).

This patch also updates trust-anchors to include system and user
certificates.

Bug: 119501137
Test: manual
Change-Id: I17a3989a5a39e57f3dff4d63a8722bd846d0fe47